+++ CHANGELOG.md
@@ -5,9 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- New lifecycle hook: [`beforeExit`](https://rize.js.org/api/interfaces/_index_.rizeoptions.html#beforeexit)
 ### Changed
 - Improved error stack trace
-- New lifecycle hook: [`beforeExit`](https://rize.js.org/api/interfaces/_index_.rizeoptions.html#beforeexit)
 
 ## [0.3.0] - 2018-03-08
 ### Added